feat: 产品增加views字段

+++ app/Http/Controllers/V1/Admin/HotelController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\V1\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Services\HotelService;
+use App\Utils\CodeResponse;
 use App\Utils\Inputs\Admin\HotelPageInput;
 use App\Utils\Inputs\HotelInput;
 
@@ -15,25 +16,19 @@ class HotelController extends Controller
     {
         /** @var HotelPageInput $input */
         $input = HotelPageInput::new();
-        $columns = [
-            'id',
-            'cover',
-            'name',
-            'grade',
-            'category_id',
-            'price',
-            'score',
-            'created_at',
-            'updated_at'
-        ];
-        $page = HotelService::getInstance()->getAdminHotelPage($input, $columns);
+        $page = HotelService::getInstance()->getAdminHotelPage($input);
         return $this->successPaginate($page);
     }
 
     public function detail()
     {
         $id = $this->verifyRequiredId('id');
+
         $hotel = HotelService::getInstance()->getHotelById($id);
+        if (is_null($hotel)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前酒店不存在');
+        }
+
         return $this->success($hotel);
     }
 
@@ -52,8 +47,27 @@ class HotelController extends Controller
         $input = HotelInput::new();
 
         $hotel = HotelService::getInstance()->getHotelById($id);
+        if (is_null($hotel)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前酒店不存在');
+        }
 
         HotelService::getInstance()->updateHotel($hotel, $input);
+
+        return $this->success();
+    }
+
+    public function editViews()
+    {
+        $id = $this->verifyRequiredId('id');
+        $views = $this->verifyRequiredInteger('views');
+
+        $hotel = HotelService::getInstance()->getHotelById($id);
+        if (is_null($hotel)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前酒店不存在');
+        }
+
+        $hotel->views = $views;
+        $hotel->save();
 
         return $this->success();
     }
@@ -63,6 +77,9 @@ class HotelController extends Controller
         $id = $this->verifyRequiredId('id');
 
         $hotel = HotelService::getInstance()->getHotelById($id);
+        if (is_null($hotel)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前酒店不存在');
+        }
         $hotel->delete();
 
         return $this->success();
