+++ app/Http/Controllers/V1/Admin/RestaurantController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\Restaurant;
 use App\Services\RestaurantCategoryService;
 use App\Services\RestaurantService;
+use App\Utils\CodeResponse;
 use App\Utils\Inputs\Admin\RestaurantPageInput;
 use App\Utils\Inputs\RestaurantInput;
 
@@ -17,23 +18,19 @@ class RestaurantController extends Controller
     {
         /** @var RestaurantPageInput $input */
         $input = RestaurantPageInput::new();
-        $columns = [
-            'id',
-            'cover',
-            'name',
-            'category_id',
-            'score',
-            'created_at',
-            'updated_at'
-        ];
-        $list = RestaurantService::getInstance()->getAdminRestaurantList($input, $columns);
+        $list = RestaurantService::getInstance()->getAdminRestaurantList($input);
         return $this->successPaginate($list);
     }
 
     public function detail()
     {
         $id = $this->verifyRequiredId('id');
+
         $restaurant = RestaurantService::getInstance()->getRestaurantById($id);
+        if (is_null($restaurant)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前餐馆不存在');
+        }
+
         return $this->success($restaurant);
     }
 
@@ -51,15 +48,41 @@ class RestaurantController extends Controller
         $id = $this->verifyRequiredId('id');
         /** @var RestaurantInput $input */
         $input = RestaurantInput::new();
+
         $restaurant = RestaurantService::getInstance()->getRestaurantById($id);
+        if (is_null($restaurant)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前餐馆不存在');
+        }
+
         RestaurantService::getInstance()->updateRestaurant($restaurant, $input);
+        return $this->success();
+    }
+
+    public function editViews()
+    {
+        $id = $this->verifyRequiredId('id');
+        $views = $this->verifyRequiredInteger('views');
+
+        $restaurant = RestaurantService::getInstance()->getRestaurantById($id);
+        if (is_null($restaurant)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前餐馆不存在');
+        }
+
+        $restaurant->views = $views;
+        $restaurant->save();
+
         return $this->success();
     }
 
     public function delete()
     {
         $id = $this->verifyRequiredId('id');
+
         $restaurant = RestaurantService::getInstance()->getRestaurantById($id);
+        if (is_null($restaurant)) {
+            return $this->fail(CodeResponse::NOT_FOUND, '当前餐馆不存在');
+        }
+
         $restaurant->delete();
         return $this->success();
     }
